\name{GenAlgVarSelection-package}
\alias{GenAlgVarSelection-package}
\alias{GenAlgVarSelection}
\docType{package}
\title{
GenAlgVarSelecton
}
\description{
This package utilizes the genetic algorithm to conduct variable selection for linear models and GLMs.
}
\details{
\tabular{ll}{
Package: \tab GenAlgVarSelection\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-12-09\cr
License: \tab What license is it under?\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Joy Hou, Kevin Li, Greta Olesen

Maintainer: Kevin Li <kbl4ew@berkeley.edu>
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}

